made BlockMachine fill cells using internal fluid handler instead of external fluid handler

+++ common/src/main/java/muramasa/antimatter/capability/machine/MachineFluidHandler.java
@@ -106,7 +106,7 @@ public class MachineFluidHandler<T extends BlockEntityMachine<T>> extends FluidH
         return true;
     }
 
-    protected void tryFillCell(int slot, long maxFill) {
+    protected void tryFillCell(int slot, int maxFill) {
         if (tile.itemHandler.map(MachineItemHandler::getCellCount).orElse(0) > 0) {
             fillCell(slot, maxFill);
         }
